hw-5 fixed minor stuff, counter yet to fix
