+++ vite-react/src/components/Counter/counter-reviews.jsx
@@ -1,4 +1,4 @@
-import { Counter } from "./counter";
+import { Counter } from './counter';
 
 export const ReviewsCounter = ({ increment, decrement, count }) => {
   return (
