--- vite-react/src/components/ProgressBar/scroll-progress.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import './progbar.module.css';
-
-export const ScrollBarProgress = ({ progress }) => {
-  return <div className='progress-fill' style={{ width: `${progress}%` }}></div>
-};
