+++ vite-react/src/constants/config.jsx
@@ -3,7 +3,6 @@ export const MAX_COUNT = 5;
 
 export const SET_NAME_ACTION = 'SET_NAME_ACTION';
 export const SET_TEXT_ACTION = 'SET_TEXT_ACTION';
-// export const SET_REVIEW_ACTION = 'SET_REVIEW_ACTION';
 export const SET_RATING_INCREMENT = 'SET_RATING_INCREMENT';
 export const SET_RATING_DECREMENT = 'SET_RATING_DECREMENT';
 export const CLEAR_FORM = 'CLEAR_FORM';
