+++ vite-react/src/hooks/use-counter.js
@@ -1,19 +1,20 @@
-import { useState, useCallback } from 'react';
-import { MIN_COUNT, MAX_COUNT } from '../../constants/config';
+// import { useState, useCallback } from 'react';
+import { useState } from 'react';
+import { MIN_COUNT, MAX_COUNT } from '../constants/config';
 
 export const useCounter = () => {
   const [value, setCount] = useState(MIN_COUNT);
 
-  const decrement = useCallback(
-    () => setCount((prevCount) => Math.min(prevCount - 1, MIN_COUNT)),
-    []
-  );
-
-  const increment = useCallback(
-    () => setCount((prevCount) => Math.min(prevCount + 1, MAX_COUNT)),
-    []
-  );
-
+    const increment = () => {
+        if (value < MAX_COUNT) {
+            setCount(value + 1);
+        }
+    };
+    const decrement = () => {
+        if (value > MIN_COUNT) {
+            setCount(value - 1);
+        }
+    };
 
   return {
     value,
@@ -21,4 +22,3 @@ export const useCounter = () => {
     decrement,
   };
 };
-
