+++ vite-react/src/reducers/reducer.js
@@ -12,12 +12,9 @@ import {
 export const reducer = (state, { type, payload }) => {
   switch (type) {
     case SET_NAME_ACTION:
-    //   ValidityState(name);
       return { ...state, name: payload };
     case SET_TEXT_ACTION:
       return { ...state, text: payload };
-    // case SET_REVIEW_ACTION:
-    //   return { ...state, review: payload };
     case SET_RATING_INCREMENT:
       return {
         ...state,
